requirements.txt
apturl==0.5.2
asgiref==3.6.0
attrs==22.2.0
backports.zoneinfo==0.2.1
bcrypt==3.1.7
black==23.7.0
blinker==1.4
Brlapi==0.7.0
certifi==2019.11.28
chardet==3.0.4
charset-normalizer==3.1.0
click==8.1.6
colorama==0.4.3
command-not-found==0.3
coverage==7.2.7
cryptography==2.8
cupshelpers==1.0
dbus-python==1.2.16
defer==1.0.6
distro==1.8.0
distro-info===0.23ubuntu1
Django==4.1.7
django-js-asset==2.1.0
django-mptt==0.14.0
djangorestframework==3.14.0
docker==6.0.1
docker-compose==1.29.2
dockerpty==0.4.1
docopt==0.6.2
drf-spectacular==0.26.3
drf-spectacular-sidecar==2023.7.1
duplicity==0.8.12.0
entrypoints==0.3
exceptiongroup==1.1.2
fasteners==0.14.1
flake8==6.0.0
future==0.18.2
httplib2==0.14.0
idna==2.8
inflection==0.5.1
iniconfig==2.0.0
jsonschema==3.2.0
keyring==18.0.1
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
lockfile==0.12.2
louis==3.12.0
macaroonbakery==1.3.1
Mako==1.1.0
MarkupSafe==1.1.0
mccabe==0.7.0
meld==3.20.2
monotonic==1.5
mypy-extensions==1.0.0
netifaces==0.10.4
oauthlib==3.1.0
olefile==0.46
packaging==23.0
paramiko==2.6.0
pathspec==0.11.1
pexpect==4.6.0
Pillow==7.0.0
platformdirs==3.9.1
pluggy==1.2.0
protobuf==3.6.1
pycairo==1.16.2
pycodestyle==2.10.0
pycups==1.9.73
pyflakes==3.0.1
PyGObject==3.36.0
PyJWT==1.7.1
pymacaroons==0.13.0
PyNaCl==1.3.0
pyRFC3339==1.1
pyrsistent==0.19.3
pytest==7.4.0
pytest-cov==4.1.0
pytest-django==4.5.2
python-apt==2.0.1+ubuntu0.20.4.1
python-dateutil==2.7.3
python-debian===0.1.36ubuntu1
python-dotenv==0.21.1
pytz==2019.3
pyxdg==0.26
PyYAML==5.3.1
reportlab==3.5.34
requests==2.31.0
requests-unixsocket==0.2.0
SecretStorage==2.3.1
simplejson==3.16.0
six==1.14.0
sqlparse==0.4.3
systemd-python==234
texttable==1.6.7
tomli==2.0.1
typing-extensions==4.7.1
ubuntu-advantage-tools==8001
ubuntu-drivers-common==0.0.0
ufw==0.36
unattended-upgrades==0.1
uritemplate==4.1.1
urllib3==1.26.16
usb-creator==0.3.7
wadllib==1.3.3
websocket-client==0.59.0
whitenoise==6.5.0
xkit==0.0.0
